(LUN-1147) Links for pages on other sites are build correctly
If the page_link contained a page from a different site than the current site, the resulting URL was messed up, containing the absolute URL of the page and the domain of the current site. Fixed this by adding the correct domain name for a page's site that is different than the current one.